dag: make DAG.addNode() responsible for adding/not adding a node.
Instead of having very similar logic in every Make*Node() factory
function, shrink those down to simple wrappers around dag.addNode().
Now the check-if-node-exists logic is in one place, addNode(). The
downside is that we will occasionally construct Node objects that are
not added to the DAG, ie. left to the garbage collector to clean up.