dsl: remove support for assigning globals at top level
The idea of phase-local variables is dumb. What we really want is
variables local to a particular build script, ie. not shared with
other scripts included in the same build. Since we have no support for
hierarchical builds yet, it's a bit premature to worry about how to
specify a variable as global. And making it implicit, e.g. "variables
declared at top level are global", seems too sneaky to me. Explicit is
better than implicit.

However, when we do figure out an appropriate syntax for assigning
global variables, reverting this change will give a good starting
point for implementing the new syntax.