types: generalize adding FuObjects, so we can add anything to a filefinder
In particular, stuff like
    src = <*.c> + "extra/foo.c"
will now work. In particular, src will be a 2-element list:
    [<*.c>, "extra/foo.c"]
where the wildcard is expanded when the list is expanded.

This was done by tweaking FuList to be a bit more lazy. In particular,
its Expand() and CommandString() methods now flatten the list in case
of list members that expand to more stuff -- like glob matches.
For example, expanding [a, b, [c, d], <*.c>] will return a list like
[a, b, c, d, "foo.c", "bar.c", "baz.c"].

This gets rid of FuFinderList too: hooray! one less type to worry about.
Simple is better than complex.