dag: make every Node implement FuObject as well.
The immediate rationale is so I can include Nodes in a FuList, which
will help the shift from Action to BuildRule being front-and-centre
(next patch).

Long term, this will let us expose and manipulate node objects (or
lists of them, whatever) in the build script, which is an important
feature.