dag,runtime: put build rules in charge, out in front of actions
Motivated by the need to define rule-local variables before running
the rule's action. Turns out that the best way to do this is
disconnect Nodes from Actions, instead connecting each (target) Node
to the BuildRule that knows how to build it. Then building a node
boils down to calling Execute() on that *build rule*, which has lots
of side benefits:

- the build rule knows all targets, so we can correctly set the
  node state for all of them
- we can set $TARGETS and $SOURCES from the build rule, not the Node
- lets us move action.go from dag to runtime, which is where it really
  wants to be: now the build *logic* is in dag, but the dirty work of
  executing stuff is in runtime