dag: make it so Nodes know their own ID
This is the start of detaching build.go from its overly intimate
relationship with the DAG. I want that code to move to a separate
package, and that means it should only interact with DAG and Node, and
not know anything about node IDs.

The cost is that DAG.Rebuild() is now a destructive operation: it has
to detach nodes from the old DAG by setting their ID to -1, and then
it trashes the old DAG since it's no longer coherent.

Added DAG.verify() and DAG.copy() methods to help test code.