dag: make it possible to define methods of NodeSet
On the surface this is simple: make NodeSet an alias for bit.Set
rather than *bit.Set. Of course it was a little more complex than
that, since it required every use of NodeSet to become *NodeSet
instead. But the upside is that other packages are now shielded a
little bit more from *bit.Set, and the dag package hides behind the
NodeSet abstraction a little more.

Also added the first method of NodeSet: String(). This replaces
setToString() in dag_test.go.