build: don't crash when target exists but is not in the DB
This happens if the Fubsy database does not exist or is empty, even
though everything has been built recently. That's not an everyday use
case, but we have to handle it more gracefully than an unexpected
panic! (In particular, we have to rebuild the world because we don't
know what changed -- same as SCons does.)