runtime,dag: drop the DAG rebuild step
It looks like we can get away without it: just leave FinderNodes as-is
in the graph, and expand wildcards when expanding actions, i.e. as
late as possible.

This means I had to implement Signature() and Exists() in FinderNode.
As soon as one of them *or* Expand() is called, we'll expand the
wildcard -- so it'll probably happen a little sooner than running
actions, but not much.

This also meant we could get rid of DAG.FindRelevantNodes().