dag,types: formally distinguish DAG expansion from action expansion
There are two motivating use cases:

  1) filefinders with variable references, e.g.
       src = <$dir/**/*.c>
     need to expand $dir before we walk the DAG, in fact before
     matching targets -- the user isn't going to ask for literal
     '$dir', but for the directory that $dir expands to. But we
     don't want to expand the wildcard at the same time, since that
     should happen later.

  2) a build rule like
       "$dir/myapp": "$dir/main.c" {
           ...
       }
     again we want to expand $dir in the two FileNodes created there,
     but we don't want to expand all wildcards at the same time --
     so we need to distinguish the two types of expansion

The solution is to distinguish expansion that happens in the DAG from
expansion that happens in an action. DAG expansion manifests as
Node.NodeExpand(), so it only has to be implemented by Node types --
which makes sense, since it only affects nodes in the DAG. It's
responsible for turning strings like "$dir/**/*.c" or "$dir/myapp"
into something that can be used for target matching:
"src/myapp/**/*.c" or "src/myapp/myapp", say. DAG expansion happens
after the main phase completes, but before the build starts running.
The other expansion is action expansion, which manifests as
FuObject.ActionExpand(), i.e. every Fubsy type has to support it. It
also does variable expansion, because command strings often contain
variables ("$CC $CFLAGS ..."), but they'll never appear as DAG nodes.
But it also takes care of expanding wildcards -- so this is where a
FinderNode will be replaced by the list of matching filenames.