runtime: make assign(), evaluate*() methods of *Runtime
They used to be standalone functions that were ignorant of Runtime.
Then they were standalone functions that all took a *Runtime as first
argument, and often used it directly. So making them methods of the
*Runtime is not a big leap.

This means that methods of *Runtime are split between two source
files: runtime.go and execute.go. Hmmmmm. Not sure how I feel about
that.