types: change so String() returns quoted values; introduce ValueString()
At least for FuString, ValueString() replaces the old String(): it
returns a filename that can be opened, a command that can be run, etc.
String() returns a debugging-friendly value, using Fubsy syntax to
explain a value where possible. The rationale is that debug messages
(Printf() etc.) will use String() by default, and code that needs the
"true" value to interact with the system should know what it's doing
and use ValueString().