runtime: expand function arguments when appropriate (from an action)
Actions should only run in the build phase, so it's fine to do
ActionExpand() from them. And it's consistent: CommandAction was
already responsible for expanding the command string, so it makes
sense for FunctionCallAction to be responsible for expanding function
args.

Fixes the regression introduced by fa1ec190cb9a.