runtime: add build() builtin for defining build rules
Of course you can already define build rules with the special syntax
for them. This is really intended for plugins, which can't be written
in Fubsy syntax; they'll need programmatic access to define build
rules. Supporting it in the DSL is optional, but seems like a nice
touch for completeness.