build scripts: factor out configure.sh; start using .build directory
configure.sh is responsible for probing the system, downloading
dependencies, and building dependencies. It's OK to be slow, because
it should run rarely. build.sh builds Fubsy, so it should be fast.

Both of these scripts should write build products to .build/1, the
"stage 1" build directory. Stage 2 will be for the self-hosted build,
i.e. using the "fubsy" binary from stage 1 to build itself. Not quite
there yet!